remove extra function

+++ src/mjs/api/list/delete.mjs
@@ -1,28 +1,32 @@
-import { API_AUCTION, API_KEY } from "../constants.mjs";
 
-export async function deleteAuction(id, authToken) {
+import { API_AUCTION } from "../constants.mjs";
+import { prepareAuthHeaders } from "../../utilities/authUtils.mjs"; 
+
+export async function deleteAuction(id) {
+  if (!id) {
+    throw new Error("Invalid auction ID");
+  }
+
+  const url = `${API_AUCTION}/${id}`;
+  const headers = prepareAuthHeaders(); 
+
   try {
-    const url = `${API_AUCTION}/${id}`;
-
-    const headers = {
-      "Content-Type": "application/json",
-      Authorization: `Bearer ${authToken}`,
-      "X-Noroff-API-Key": API_KEY,
-    };
-
     const response = await fetch(url, {
       method: "DELETE",
-      headers: headers,
+      headers,
     });
 
     if (!response.ok) {
       const errorData = await response.json();
-      throw new Error(errorData.message || "Failed to delete auction listing.");
+      const errorMessage = errorData.message || "Failed to delete auction listing.";
+      console.error(`Error ${response.status}: ${errorMessage}`);
+      throw new Error(errorMessage); 
     }
 
-    return { success: true, message: "Post deleted successfully" };
+    
+    return { success: true, message: "Auction deleted successfully." };
   } catch (error) {
     console.error("Error deleting auction listing:", error);
-    throw error;
+    throw error; 
   }
 }
